+++ packages/espressocash_app/lib/features/outgoing_direct_payments/screens/network_picker_screen.dart
@@ -47,8 +47,10 @@ class NetworkPickerRoute extends GoRouteData {
       );
 }
 
+// TODO(KB): refactor to class
 typedef NetworkPickerParams = ({
   Blockchain? initial,
+  // ignore: avoid-function-type-in-records, refactor later
   ValueSetter<Blockchain> onSubmitted,
 });