+++ packages/espressocash_app/lib/features/outgoing_dln_payments/services/confirm_payment_bloc.dart
@@ -154,6 +154,7 @@ extension ConfirmPaymentExt on ConfirmPaymentState {
       quote?.receiverAmount ??
       const CryptoAmount(value: 0, cryptoCurrency: Currency.usdc);
 
+  @useResult
   Either<CreateOrderException, PaymentQuote> validate(
     CryptoAmount usdcBalance,
   ) {